[chore] #103 Move custom component in sub folder
