+++ common/design/src/main/java/com/cmc/curtaincall/common/design/component/custom/RatingBar.kt
@@ -1,4 +1,4 @@
-package com.cmc.curtaincall.common.design.component
+package com.cmc.curtaincall.common.design.component.custom
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
